pyproject.toml
[project]
name = "semantic-weight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
]
